+++ up.php
@@ -3,14 +3,14 @@
 /**
  *
  * @package plg_UP for Joomla! 3.0+
- * @version $Id: up.php 2017-07-02 $
+ * @version $Id: up.php 2025-11-01 $
  * @author Lomart
- * @copyright (c) 2017 Lomart
+ * @copyright (c) 2025 Lomart
  * @license   <a href="http://www.gnu.org/licenses/gpl-3.0.html" target="_blank">GNU/GPLv3</a>
  *
  * */
 /*
-v5.3.3 : php 8.4 compatibility
+v5.3.3 : php 8.4/8.5 compatibility
 */
 
 // namespace up;
@@ -548,7 +548,6 @@ class plgContentUP extends CMSPlugin
             }
 
             $response = curl_exec($curl);
-            curl_close($curl);
             return $response;
         } catch (\RuntimeException $e) {
             return null;