+++ changelog.md
@@ -6,6 +6,7 @@
 
 #### Modifications internes des actions
 
+- compatibilité PHP 8.4 : définition des varaibles dans upAction.php
 - jcontent_by_categories : compatibilité Joomla 6 : remplacement de getInstance()
 - jcontent_by_subcat : compatibilité Joomla 6 : remplacement de getInstance()
 - jcontent_in_content : compatibilité Joomla 6 : remplacement de getInstance()