+++ changelog.md
@@ -4,6 +4,7 @@
 
 - package sans les actions, qui seront téléchargées quand nécessaires à partir de github
 - vérification des versions des actions et rechargement depuis github si différence
+- nouveau paramètre : vérification des actions sur github
 
 #### Modifications internes des actions