+++ changelog.md
@@ -8,7 +8,8 @@
 
 #### Modifications internes des actions
 
-- compatibilité PHP 8.4 : définition des varaibles dans upAction.php
+- compatibilité PHP 8.4 : définition des variables dans upAction.php
+- compatibilité PHP 8.5 : curl_close deprecated
 - jcontent_by_categories : compatibilité Joomla 6 : remplacement de getInstance()
 - jcontent_by_subcat : compatibilité Joomla 6 : remplacement de getInstance()
 - jcontent_in_content : compatibilité Joomla 6 : remplacement de getInstance()