+++ changelog.md
@@ -1,4 +1,4 @@
-## 22/10/2025 - version 5.3.3
+## 01/11/2025 - version 5.3.3
 
 #### Modifications actions